:recycle: Refactor networking related code into separate modules
